Rename infra directories
